EGYBEST/requirements.txt
requests
bs4
pySmartDL
selenium
helium
msedge-selenium-tools